Skip load-and-authorize in handle_params after mount; drop LV <0.20
LiveView 0.19 and 0.18 were found to have issues when using streams for loading and authorizing a collection, and for compatibility with 0.20 (and of course for extra performance to avoid the double query) the handle_params hook is skipped if we're in the mounting phase. In LiveView >= 1.0.2, since https://github.com/phoenixframework/phoenix_live_view/pull/3599, display would be accurate anyway because of deduplication, but it was good to have it implemented for speed anyway.